+++ example/android/app/src/main/kotlin/braze/com/brazepluginexample/EmbeddingV1Activity.kt
@@ -9,17 +9,17 @@ import io.flutter.app.FlutterActivity
 import io.flutter.view.FlutterMain
 
 import android.view.View
-import com.appboy.Appboy
-import com.appboy.AppboyLifecycleCallbackListener
-import com.appboy.events.ContentCardsUpdatedEvent
 import com.appboy.events.IEventSubscriber
-import com.appboy.models.IInAppMessage
-import com.appboy.models.MessageButton
 import com.appboy.models.cards.Card
-import com.appboy.ui.inappmessage.AppboyInAppMessageManager
-import com.appboy.ui.inappmessage.InAppMessageCloser
-import com.appboy.ui.inappmessage.InAppMessageOperation
-import com.appboy.ui.inappmessage.listeners.IInAppMessageManagerListener
+import com.braze.Braze
+import com.braze.BrazeActivityLifecycleCallbackListener
+import com.braze.events.ContentCardsUpdatedEvent
+import com.braze.models.inappmessage.IInAppMessage
+import com.braze.models.inappmessage.MessageButton
+import com.braze.ui.inappmessage.BrazeInAppMessageManager
+import com.braze.ui.inappmessage.InAppMessageCloser
+import com.braze.ui.inappmessage.InAppMessageOperation
+import com.braze.ui.inappmessage.listeners.IInAppMessageManagerListener
 
 class EmbeddingV1Activity : FlutterActivity() {
   private var mContentCardsUpdatedSubscriber: IEventSubscriber<ContentCardsUpdatedEvent>? = null
@@ -31,29 +31,25 @@ class EmbeddingV1Activity : FlutterActivity() {
     BrazePlugin.registerWith(registrarFor("com.braze.brazeplugin.BrazePlugin"))
     IntegrationTestPlugin.registerWith(registrarFor("dev.flutter.plugins.integration_test.IntegrationTestPlugin"))
 
-    this.getApplication().registerActivityLifecycleCallbacks(AppboyLifecycleCallbackListener())
+    this.getApplication().registerActivityLifecycleCallbacks(BrazeActivityLifecycleCallbackListener())
     this.subscribeToContentCardsUpdatedEvent()
 
-    AppboyInAppMessageManager.getInstance().setCustomInAppMessageManagerListener(
+    BrazeInAppMessageManager.getInstance().setCustomInAppMessageManagerListener(
             MainActivity.BrazeInAppMessageManagerListener())
-    Appboy.getInstance(this).logCustomEvent("flutter_sample_opened_v1")
+    Braze.getInstance(this).logCustomEvent("flutter_sample_opened_v1")
   }
 
   private fun subscribeToContentCardsUpdatedEvent() {
-    Appboy.getInstance(this).removeSingleSubscription(mContentCardsUpdatedSubscriber, ContentCardsUpdatedEvent::class.java)
+    Braze.getInstance(this).removeSingleSubscription(mContentCardsUpdatedSubscriber, ContentCardsUpdatedEvent::class.java)
     mContentCardsUpdatedSubscriber = IEventSubscriber { event ->
       val allCards = event.allCards as ArrayList<Card>
       BrazePlugin.processContentCards(allCards)
     }
-    Appboy.getInstance(this).subscribeToContentCardsUpdates(mContentCardsUpdatedSubscriber)
-    Appboy.getInstance(this).requestContentCardsRefresh(true)
+    Braze.getInstance(this).subscribeToContentCardsUpdates(mContentCardsUpdatedSubscriber)
+    Braze.getInstance(this).requestContentCardsRefresh(true)
   }
 
   class BrazeInAppMessageManagerListener() : IInAppMessageManagerListener {
-    override fun onInAppMessageReceived(inAppMessage: IInAppMessage): Boolean {
-      return false
-    }
-
     override fun beforeInAppMessageDisplayed(inAppMessage: IInAppMessage): InAppMessageOperation {
       BrazePlugin.processInAppMessage(inAppMessage)
       return InAppMessageOperation.DISPLAY_NOW