+++ example/integration_test/braze_plugin_integration_test.dart
@@ -6,9 +6,9 @@ void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
   testWidgets('Can initialize plugin', (WidgetTester tester) async {
-   BrazePlugin plugin = BrazePlugin();
-   Future<String> installTrackingId = plugin.getInstallTrackingId();
-   expect(plugin, isNotNull);
-   expect(installTrackingId, isNotNull);
+    BrazePlugin plugin = BrazePlugin();
+    Future<String> installTrackingId = plugin.getInstallTrackingId();
+    expect(plugin, isNotNull);
+    expect(installTrackingId, isNotNull);
   });
 }